backend/requirements.txt
# backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
Pillow==10.3.0
pytesseract==0.3.10
PyPDF2==3.0.1
transformers==4.41.1
google-cloud-firestore==2.16.0
langchain==0.1.20
langchain-google-vertexai==0.1.6
langchain-core==0.1.52
langchain-community==0.0.38
langchain-huggingface==0.0.3
faiss-cpu==1.8.0
python-multipart==0.0.9
passlib[bcrypt]==1.7.4